// src/_db/drizzle/enum/inventory-movement-type.enum.ts
export enum InventoryMovementTypeEnum {
  INITIAL_STOCK = 'INITIAL_STOCK',
  RESTOCK = 'RESTOCK',
  ORDER_RESERVED = 'ORDER_RESERVED',
  ORDER_FULFILLED = 'ORDER_FULFILLED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  CUSTOMER_RETURN = 'CUSTOMER_RETURN',
  DAMAGED = 'DAMAGED',
  LOST = 'LOST',
  ADJUSTMENT = 'ADJUSTMENT',
  TRANSFER_OUT = 'TRANSFER_OUT',
  TRANSFER_IN = 'TRANSFER_IN',
}